+++ src/pages/scientificServices/pipelines/tabs/run/inputs/file/GcsFileInput.tsx
@@ -1,165 +1,10 @@
-import { Icon, Spinner } from '@terra-ui-packages/components';
 import React, { ReactNode, useState } from 'react';
-import { ClipboardButton } from 'src/components/ClipboardButton';
 import { PipelineInput } from 'src/libs/ajax/teaspoons/teaspoons-models';
 import colors from 'src/libs/colors';
-import { getConfig } from 'src/libs/config';
 import { getTerraUser } from 'src/libs/state';
-import { DocsKey, ZendeskLink } from 'src/pages/scientificServices/pipelines/common/zendeskUtils';
-import { BucketConsoleLink } from 'src/pages/scientificServices/pipelines/tabs/run/inputs/file/BucketConsoleLink';
+import { SharingInstructions } from 'src/pages/scientificServices/pipelines/common/SharingInstructions';
 import { GCS_PATH_VALIDATION_REGEX } from 'src/pages/scientificServices/pipelines/utils/upload-utils';
 import { useProxyGroup } from 'src/profile/personal-info/useProxyGroup';
-
-const renderProxyGroupContent = (isLoading: boolean, proxyGroupEmail: string | null) => {
-  if (isLoading) {
-    return (
-      <div style={{ display: 'flex', alignItems: 'center', flex: 1 }}>
-        <Spinner size={16} />
-        <span style={{ color: '#666' }}>Loading proxy group...</span>
-      </div>
-    );
-  }
-
-  if (proxyGroupEmail) {
-    return (
-      <>
-        <code
-          style={{
-            flex: 1,
-            padding: '0.5rem',
-            backgroundColor: '#fff',
-            border: '1px solid #e0e0e0',
-            borderRadius: '4px',
-            fontSize: '12px',
-            overflow: 'hidden',
-            textOverflow: 'ellipsis',
-          }}
-        >
-          {proxyGroupEmail}
-        </code>
-        <ClipboardButton text={proxyGroupEmail} />
-      </>
-    );
-  }
-
-  return (
-    <div style={{ color: '#d00', display: 'flex', alignItems: 'center' }}>
-      <Icon icon='warning-standard' size={16} style={{ color: '#d00', marginRight: '0.25rem' }} />
-      Failed to load proxy group information. Please refresh the page.
-    </div>
-  );
-};
-
-interface SharingInstructionsProps {
-  isExpanded: boolean;
-  onToggleExpand: () => void;
-  proxyGroupEmail: string | null;
-  isLoadingProxyGroup: boolean;
-  cloudPath?: string;
-}
-
-const SharingInstructions: React.FC<SharingInstructionsProps> = ({
-  isExpanded,
-  onToggleExpand,
-  proxyGroupEmail,
-  isLoadingProxyGroup,
-  cloudPath,
-}) => {
-  const serviceAccountEmail = getTeaspoonsServiceAccountEmail();
-
-  return (
-    <>
-      <div style={{ marginTop: '0.5rem' }}>
-        <button
-          type='button'
-          onClick={onToggleExpand}
-          style={{
-            background: 'none',
-            border: 'none',
-            color: '#46A3E9',
-            cursor: 'pointer',
-            fontSize: '14px',
-            textDecoration: 'underline',
-            padding: 0,
-            display: 'flex',
-            alignItems: 'center',
-            gap: '0.25rem',
-          }}
-        >
-          <Icon icon={isExpanded ? 'angle-down' : 'angle-right'} size={16} style={{ flexShrink: 0 }} />
-          View sharing instructions
-        </button>
-      </div>
-      {isExpanded && (
-        <div
-          style={{
-            marginTop: '0.5rem',
-            padding: '0.75rem',
-            backgroundColor: '#f5f5f5',
-            borderRadius: '4px',
-            border: '1px solid #e0e0e0',
-          }}
-        >
-          <div style={{ fontSize: '14px', color: '#333', marginBottom: '1rem' }}>
-            To ensure that your input file can be properly accessed by Broad Scientific Services, please share your
-            input file with the following accounts:
-          </div>
-          <div style={{ fontSize: '14px', fontWeight: 600, marginBottom: '0.5rem' }}>Service account</div>
-          <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem' }}>
-            <code
-              style={{
-                flex: 1,
-                padding: '0.5rem',
-                backgroundColor: '#fff',
-                border: '1px solid #e0e0e0',
-                borderRadius: '4px',
-                fontSize: '12px',
-                overflow: 'hidden',
-                textOverflow: 'ellipsis',
-              }}
-            >
-              {serviceAccountEmail}
-            </code>
-            <ClipboardButton text={serviceAccountEmail} />
-          </div>
-          <div
-            style={{
-              fontSize: '14px',
-              marginBottom: '0.5rem',
-              marginTop: '1rem',
-              fontWeight: 600,
-            }}
-          >
-            Your proxy group
-          </div>
-          <div style={{ display: 'flex', alignItems: 'center', gap: '0.5rem', marginBottom: '1rem' }}>
-            {renderProxyGroupContent(isLoadingProxyGroup, proxyGroupEmail)}
-          </div>
-          <div
-            style={{
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'space-between',
-              paddingTop: '0.5rem',
-              gap: '0.5rem',
-            }}
-          >
-            <div style={{ fontSize: '14px', color: '#666' }}>
-              <ZendeskLink docsKey={DocsKey.CLOUD_INPUTS}>Learn more</ZendeskLink> about file sharing requirements.
-            </div>
-            <ClipboardButton
-              text={`${serviceAccountEmail}, ${proxyGroupEmail || ''}`}
-              style={{ display: 'flex', alignItems: 'center', flexShrink: 0 }}
-            >
-              Copy all
-            </ClipboardButton>
-          </div>
-          <BucketConsoleLink cloudPath={cloudPath} />
-        </div>
-      )}
-    </>
-  );
-};
 
 interface GcsFileInputProps {
   input: PipelineInput;
@@ -269,7 +114,7 @@ export const GcsFileInput: React.FC<GcsFileInputProps> = ({
               onChange={(e) => onSharingConfirmationChange?.(e.target.checked)}
               style={{ marginTop: '0.25rem', cursor: 'pointer' }}
             />
-            <span style={{ fontSize: '14px', flex: 1 }}>
+            <span style={{ fontSize: '14px', flex: 1, marginBottom: '0.25rem' }}>
               I have shared this file with Broad Scientific Services.{' '}
               <span style={{ color: colors.danger(), fontWeight: 'bold' }}>*</span>
             </span>
@@ -285,12 +130,4 @@ export const GcsFileInput: React.FC<GcsFileInputProps> = ({
       </div>
     </div>
   );
-};
-
-const getTeaspoonsServiceAccountEmail = (): string => {
-  if (getConfig().isProd) {
-    return 'broad-scientific-services@firecloud.org';
-  }
-
-  return 'broad-scientific-services@dev.test.firecloud.org';
 };
